requirements.txt
telethon
datetime
asyncio
requests
python-dotenv
telebot
uuid
